Changed db backend from mongodb/mongoengine to sqlite/sqlalchemy.

+++ user.py
@@ -1,30 +1,22 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+import datetime
+import os
+import time
+
 from flask import (Flask, render_template, flash,
                    url_for, jsonify, abort, make_response)
 from flask import request
 from flask import redirect
 
-import time
-import datetime
+from netaddr import IPAddress, IPNetwork
+from shortuuid import uuid
+
 
 from flask import Blueprint
 from flask.views import MethodView
-from fileexchange.models import File
-import os
-
-import struct
-import socket
-
-
-def addressInNetwork(ip, net):
-    "Is an address in a network"
-    ipaddr = struct.unpack('L', socket.inet_aton(ip))[0]
-    netaddr, bits = net.split('/')
-    netmask = struct.unpack('L',
-                            socket.inet_aton(netaddr))[0] & ((2L<<int(bits)-1) - 1)
-    return ipaddr & netmask == netmask
+from fileexchange.models import session, File, Log
 
 user = Blueprint('user', __name__, template_folder='templates')
 
@@ -34,27 +26,28 @@ class Index(MethodView):
     def get(self):
         try:
             remote = request.headers["X-Forwarded-For"]
-            if addressInNetwork(remote, "192.168.0.0/24"):
+            if IPAddress(remote) in IPNetwork("192.168.0.0/24"):
                 return render_template("index.html")
             else:
                 abort(403)
-        except:
+        except None:
             abort(403)
 
 
 class FileList(MethodView):
     def get(self):
-        files = File.objects(owner=request.headers["X-Forwarded-For"]).order_by("-uploaded")
+        files = session.query(File).filter_by(owner=request.headers["X-Forwarded-For"]).order_by("-upload_date")
         return render_template("filelist.html", files=files)
 
 
 class DeleteFile(MethodView):
     def get(self, id=None):
         try:
-            f = File.objects.get(id=id)
-            os.unlink(f.path)
-            f.delete()
-            return "ok"
+            f = session.query(File).filter_by(owner=request.headers["X-Forwarded-For"], uuid=id).first()
+            if f:
+                os.unlink(f.path)
+                f.delete()
+                return "ok"
         except:
             abort(500)
 
@@ -62,36 +55,40 @@ class DeleteFile(MethodView):
 class GetFile(MethodView):
     def get(self, id=None, filename=None):
         try:
-            f = File.objects.get(id=id)
-            File.objects(id=id).update(inc__count=1)
-            File.objects(id=id).update(push__downloadtime=datetime.datetime.now())
-            response = make_response()
-            response.headers['Cache-Control'] = 'no-cache'
-            response.headers['Content-Type'] = f.type
-            response.headers['Content-Disposition'] = 'attachment'
-            response.headers['X-Accel-Redirect'] = "/download/" + "/".join(f.path.split("/")[3:])
-            return response
+            f = session.query(File).filter_by(uuid=id).first()
+            if f:
+                log = Log(request.headers["X-Forwarded-For"], f)
+                session.add(log)
+                session.commit()
+                response = make_response()
+                response.headers['Cache-Control'] = 'no-cache'
+                response.headers['Content-Type'] = f.type
+                response.headers['Content-Disposition'] = 'attachment'
+                response.headers['X-Accel-Redirect'] = "/download/" + "/".join(f.path.split("/")[3:])
+                return response
         except:
+            session.rollback()
             abort(404)
 
 
 class Uploader(MethodView):
-
     def post(self):
         res = []
         f = File()
+        f.uuid = uuid()
         f.name = request.form["qqfile.name"]
         f.size = request.form["qqfile.size"]
-        f.name = request.form["qqfile.name"]
         f.path = request.form["qqfile.path"]
+        f.owner = request.headers["X-Forwarded-For"]
         f.type = request.form["qqfile.content_type"]
         f.md5 = request.form["qqfile.md5"]
-        f.savetime = datetime.datetime.now() + datetime.timedelta(hours=int(request.args.get("savetime")))
-        f.owner = request.headers["X-Forwarded-For"]
+        f.deletion_date = datetime.datetime.now() + datetime.timedelta(hours=int(request.args.get("savetime")))
+        f.upload_date = datetime.datetime.now()
         try:
-            f.validate()
-            f.save()
-        except None:
+            session.add(f)
+            session.commit()
+        except:
+            session.rollback()
             return make_response("{\"error\"}")
         return make_response("{\"success\": true}")
